Move form utils to module
